\alias{banknote}
\name{Swiss Banknote Data}
\docType{data}
\title{
The Swiss Banknote Data
}
\description{
This package contains measurements on 200 Swiss banknotes: 100 genuine and 100 counterfeit. The variables are length of bill, width of left edge, width of right edge , bottom margin width and top margin width. All measurements are in millimetres. The data source is noted below. This data is also available in the alr package in R. 
}

\references{
	Flury, B. and Riedwyl, H. (1988). Multivariate Statistics: A practical approach. London: Chapman and Hall.
}
\usage{data(banknote)}
\examples{
data(banknote) # Loads the brown bread data set
head(banknote) # Displays the first six rows of the brown bread data set
}
\keyword{ datasets }
\keyword{ Banknote Data }
