\name{uskewFactors}
\alias{uskewFactors}
\docType{package}
\title{
Model-Based Clustering via Mixtures of 'Unrestricted' Skew-t Factor Analyzers 
}
\description{
Contains the function uskewFA for fitting mixtures of 'unrestricted' skew-t factor analyzer models
}
\details{
\tabular{ll}{
Package: \tab uskewFactors\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-22\cr
License: \tab WGPL (>=2)\cr
}

}
\author{
Paula M. Murray, Ryan P. Browne, and Paul D. McNicholas

Maintainer: Paula M. Murray <murrapm3@math.mcmaster.ca>
}
\references{
Murray, P.M., Browne, R.P., and McNicholas, P.D. (2014), "Mixtures of 'Unrestricted' Skew-t Factor Analyzers". Arxiv preprint arXiv:1310.6224}
\keyword{ Expectation Maximization (EM) Algorithm  }
\keyword{ Clustering }
\keyword{Skew-t Distribution}
\keyword{Factor Analysis}
\keyword{Mixture Models}
\seealso{
Details, references, and examples are given under \link{uskewFA}.
}

